more overwrite logic and tests

+++ sdk/go/common/workspace/templates.go
@@ -743,11 +743,13 @@ func writeAllBytes(filename string, bytes []byte, overwrite bool, mode os.FileMo
 	} else {
 		flag = flag | os.O_EXCL
 	}
-
-	info, err := os.Stat(filename)
-	if err == nil {
-		if info.IsDir() {
-			os.Remove(filename)
+	
+	if overwrite {
+		info, err := os.Stat(filename)
+		if err == nil {
+			if info.IsDir() {
+				os.Remove(filename)
+			}
 		}
 	}
